requirements.txt
openai
langchain
pinecone-client
python-dotenv
tiktoken
huggingface_hub
streamlit